my_blog/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
defusedxml==0.7.1
Django==2.2.13
django-allauth==0.39.1
django-ckeditor==5.6.1
django-js-asset==2.0.0
django-model-utils==4.3.1
django-mptt==0.10.0
django-notifications-hq==1.5.0
django-password-reset==2.0
django-taggit==0.23.0
form==0.0.1
idna==3.4
#importlib-metadata==5.0.0
jsonfield==3.1.0
Markdown==3.4.1
oauthlib==3.2.2
Pillow==5.3.0
pycparser==2.21
Pygments==2.13.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.12
zipp==3.10.0
